refactor: rename client query package to request
