+++ clients/java/src/main/java/io/camunda/client/CamundaClient.java
@@ -80,17 +80,17 @@ import io.camunda.client.api.fetch.UserTaskGetRequest;
 import io.camunda.client.api.fetch.VariableGetRequest;
 import io.camunda.client.api.response.ActivatedJob;
 import io.camunda.client.api.response.DocumentReferenceResponse;
-import io.camunda.client.api.search.query.AdHocSubprocessActivitySearchRequest;
-import io.camunda.client.api.search.query.DecisionDefinitionSearchRequest;
-import io.camunda.client.api.search.query.DecisionInstanceSearchRequest;
-import io.camunda.client.api.search.query.DecisionRequirementsSearchRequest;
-import io.camunda.client.api.search.query.FlownodeInstanceSearchRequest;
-import io.camunda.client.api.search.query.IncidentSearchRequest;
-import io.camunda.client.api.search.query.ProcessDefinitionSearchRequest;
-import io.camunda.client.api.search.query.ProcessInstanceSearchRequest;
-import io.camunda.client.api.search.query.UserTaskSearchRequest;
-import io.camunda.client.api.search.query.UserTaskVariableSearchRequest;
-import io.camunda.client.api.search.query.VariableSearchRequest;
+import io.camunda.client.api.search.request.AdHocSubprocessActivitySearchRequest;
+import io.camunda.client.api.search.request.DecisionDefinitionSearchRequest;
+import io.camunda.client.api.search.request.DecisionInstanceSearchRequest;
+import io.camunda.client.api.search.request.DecisionRequirementsSearchRequest;
+import io.camunda.client.api.search.request.FlownodeInstanceSearchRequest;
+import io.camunda.client.api.search.request.IncidentSearchRequest;
+import io.camunda.client.api.search.request.ProcessDefinitionSearchRequest;
+import io.camunda.client.api.search.request.ProcessInstanceSearchRequest;
+import io.camunda.client.api.search.request.UserTaskSearchRequest;
+import io.camunda.client.api.search.request.UserTaskVariableSearchRequest;
+import io.camunda.client.api.search.request.VariableSearchRequest;
 import io.camunda.client.api.worker.JobClient;
 import io.camunda.client.api.worker.JobWorkerBuilderStep1;
 import io.camunda.client.impl.CamundaClientBuilderImpl;
