+++ clients/java/src/main/java/io/camunda/client/api/search/filter/UserTaskVariableFilter.java
@@ -16,7 +16,7 @@
 package io.camunda.client.api.search.filter;
 
 import io.camunda.client.api.search.filter.builder.StringProperty;
-import io.camunda.client.api.search.query.TypedSearchRequest.SearchRequestFilter;
+import io.camunda.client.api.search.request.TypedSearchRequest.SearchRequestFilter;
 import java.util.function.Consumer;
 
 public interface UserTaskVariableFilter extends SearchRequestFilter {
