+++ clients/java/src/main/java/io/camunda/client/api/search/filter/VariableValueFilter.java
@@ -15,7 +15,7 @@
  */
 package io.camunda.client.api.search.filter;
 
-import io.camunda.client.api.search.query.TypedSearchRequest.SearchRequestFilter;
+import io.camunda.client.api.search.request.TypedSearchRequest.SearchRequestFilter;
 
 public interface VariableValueFilter extends SearchRequestFilter {
