+++ clients/java/src/main/java/io/camunda/client/impl/search/query/DecisionInstanceSearchRequestImpl.java
@@ -23,8 +23,8 @@ import io.camunda.client.api.CamundaFuture;
 import io.camunda.client.api.JsonMapper;
 import io.camunda.client.api.search.SearchRequestPage;
 import io.camunda.client.api.search.filter.DecisionInstanceFilter;
-import io.camunda.client.api.search.query.DecisionInstanceSearchRequest;
-import io.camunda.client.api.search.query.FinalSearchRequestStep;
+import io.camunda.client.api.search.request.DecisionInstanceSearchRequest;
+import io.camunda.client.api.search.request.FinalSearchRequestStep;
 import io.camunda.client.api.search.response.DecisionInstance;
 import io.camunda.client.api.search.response.SearchQueryResponse;
 import io.camunda.client.api.search.sort.DecisionInstanceSort;
