+++ clients/java/src/main/java/io/camunda/client/impl/search/query/DecisionRequirementsSearchRequestImpl.java
@@ -22,8 +22,8 @@ import io.camunda.client.api.JsonMapper;
 import io.camunda.client.api.search.SearchRequestBuilders;
 import io.camunda.client.api.search.SearchRequestPage;
 import io.camunda.client.api.search.filter.DecisionRequirementsFilter;
-import io.camunda.client.api.search.query.DecisionRequirementsSearchRequest;
-import io.camunda.client.api.search.query.FinalSearchRequestStep;
+import io.camunda.client.api.search.request.DecisionRequirementsSearchRequest;
+import io.camunda.client.api.search.request.FinalSearchRequestStep;
 import io.camunda.client.api.search.response.DecisionRequirements;
 import io.camunda.client.api.search.response.SearchQueryResponse;
 import io.camunda.client.api.search.sort.DecisionRequirementsSort;
