+++ clients/java/src/main/java/io/camunda/client/impl/search/query/ProcessDefinitionSearchRequestImpl.java
@@ -23,8 +23,8 @@ import io.camunda.client.api.CamundaFuture;
 import io.camunda.client.api.JsonMapper;
 import io.camunda.client.api.search.SearchRequestPage;
 import io.camunda.client.api.search.filter.ProcessDefinitionFilter;
-import io.camunda.client.api.search.query.FinalSearchRequestStep;
-import io.camunda.client.api.search.query.ProcessDefinitionSearchRequest;
+import io.camunda.client.api.search.request.FinalSearchRequestStep;
+import io.camunda.client.api.search.request.ProcessDefinitionSearchRequest;
 import io.camunda.client.api.search.response.ProcessDefinition;
 import io.camunda.client.api.search.response.SearchQueryResponse;
 import io.camunda.client.api.search.sort.ProcessDefinitionSort;
