+++ clients/java/src/main/java/io/camunda/client/impl/search/query/UserTaskSearchRequestImpl.java
@@ -22,8 +22,8 @@ import static io.camunda.client.api.search.SearchRequestBuilders.userTaskSort;
 import io.camunda.client.api.JsonMapper;
 import io.camunda.client.api.search.SearchRequestPage;
 import io.camunda.client.api.search.filter.UserTaskFilter;
-import io.camunda.client.api.search.query.FinalSearchRequestStep;
-import io.camunda.client.api.search.query.UserTaskSearchRequest;
+import io.camunda.client.api.search.request.FinalSearchRequestStep;
+import io.camunda.client.api.search.request.UserTaskSearchRequest;
 import io.camunda.client.api.search.response.SearchQueryResponse;
 import io.camunda.client.api.search.response.UserTask;
 import io.camunda.client.api.search.sort.UserTaskSort;
