+++ clients/java/src/main/java/io/camunda/zeebe/client/api/search/query/ProcessInstanceQuery.java
@@ -15,14 +15,13 @@
  */
 package io.camunda.zeebe.client.api.search.query;
 
-import io.camunda.client.api.search.query.ProcessInstanceSearchRequest;
+import io.camunda.client.api.search.request.ProcessInstanceSearchRequest;
 import io.camunda.zeebe.client.api.search.filter.ProcessInstanceFilter;
 import io.camunda.zeebe.client.api.search.response.ProcessInstance;
 import io.camunda.zeebe.client.api.search.sort.ProcessInstanceSort;
 
 /**
- * @deprecated since 8.8 for removal in 8.9, replaced by {@link
- *     ProcessInstanceSearchRequest}
+ * @deprecated since 8.8 for removal in 8.9, replaced by {@link ProcessInstanceSearchRequest}
  */
 @Deprecated
 public interface ProcessInstanceQuery
